Move an example to test/
