+++ src/test/scala/BaseSpec.scala
@@ -1,0 +1,4 @@
+import org.scalatest.wordspec.AnyWordSpecLike
+import org.scalatest.matchers.should.Matchers
+
+abstract class BaseSpec extends AnyWordSpecLike with Matchers
